feat: implement XMLparser
parse arguments of different types
create input.cpp necessary instead of input.cxx because of CMake
todo: main, linkedcellerror

+++ src/inputReader/XMLFileReader/Args/SphereArg.cpp
@@ -1,7 +1,8 @@
 #include "SphereArg.h"
 SphereArg::SphereArg(const std::array<double, 3>& position,
-                     const std::array<double, 3>& velocity, double radius,
-                     int dimension, double distance, double mass, int type)
+                     const std::array<double, 3>& velocity, unsigned int radius,
+                     unsigned int dimension, double distance, double mass,
+                     int type)
     : position(position),
       velocity(velocity),
       radius(radius),
@@ -14,9 +15,9 @@ const std::array<double, 3>& SphereArg::getPosition() const { return position; }
 
 const std::array<double, 3>& SphereArg::getVelocity() const { return velocity; }
 
-double SphereArg::getRadius() const { return radius; }
+unsigned int SphereArg::getRadius() const { return radius; }
 
-int SphereArg::getDimension() const { return dimension; }
+unsigned int SphereArg::getDimension() const { return dimension; }
 
 double SphereArg::getDistance() const { return distance; }
